fix(tts): cloud provider variant routing, Inworld field name, test stability [SPEC_003]
Fix 3 E2E bugs: Inworld audioConfig field (encoding→audioEncoding), cloud
providers receiving Kokoro's "int8" variant instead of their own model names
(TtsSpeaker, PipelineFactory, TtsSettingsViewModel), and environment-dependent
Kokoro tests (int8→fp32 for model-not-found assertions). Updated docs with
E2E testing results and known TTS→DB logging gap.

+++ tests/DiktaMe.Core.Tests/TTS/KokoroModelManagerTests.cs
@@ -64,8 +64,8 @@ public sealed class KokoroModelManagerTests
     [Fact]
     public void IsModelDownloaded_WhenNotPresent_ReturnsFalse()
     {
-        var mgr = new KokoroModelManager("int8");
-        // Model won't be downloaded in test environment
+        // Use fp32 variant (310MB) — virtually never downloaded in dev/CI environments
+        var mgr = new KokoroModelManager("fp32");
         mgr.IsModelDownloaded.Should().BeFalse();
     }
